chore(demos): move old demos task to demos.prod
update the demos file with shared tasks, include the shared css
