+++ scripts/demos/demos.template.dev.html
@@ -9,6 +9,7 @@
 	<meta name="apple-mobile-web-app-capable" content="yes">
 
 	<link id="ionicLink" href="../css/ionic.css" rel="stylesheet">
+	<link href="../css/demos.shared.css" rel="stylesheet">
 
 	<script src="/node_modules/systemjs/dist/system.src.js"></script>
 	<script>
