+++ scripts/gulp/gulpfile.ts
@@ -1,8 +1,9 @@
 import './tasks/build';
 import './tasks/clean';
 import './tasks/default';
-// import './tasks/demos';
+import './tasks/demos';
 import './tasks/demos.dev';
+// import './tasks/demos.prod';
 import './tasks/docs';
 import './tasks/e2e';
 import './tasks/e2e.dev';
